Split pallet aleph into aleph and elections
